Add paned widgets to the natural size test for GtkAlignment and GtkSocket.
2007-07-27  Mathias Hasselmann  <mathias.hasselmann@gmx.de>

	* tests/testextendedlayout.c: Add paned widgets to
	the natural size test for GtkAlignment and GtkSocket.
	* gtk/gtksocket-x11.c: Fix some cast.

svn path=/branches/extended-layout/; revision=18553

+++ gtk/gtksocket-x11.c
@@ -164,7 +164,7 @@ _gtk_socket_windowing_get_natural_size (GtkSocket *socket)
           return;
         }
 
-      data_long = (long*) data;
+      data_long = (gint32*) data;
       priv->natural_width = MAX (1, data_long[0]);
       priv->natural_height = MAX (1, data_long[1]);
